+++ src/main/java/controllers/InventoryPresenter.java
@@ -1,0 +1,4 @@
+package controllers;
+
+public class InventoryPresenter {
+}
