+++ src/main/java/use_case/InventoryDataAccess.java
@@ -1,0 +1,12 @@
+package use_case;
+
+import entities.GroceryItem;
+
+import java.util.ArrayList;
+
+public interface InventoryDataAccess {
+
+    ArrayList<GroceryItem> read();
+
+    void update(ArrayList<GroceryItem> items);
+}
