+++ src/main/java/use_case/OrderDataAccess.java
@@ -1,0 +1,10 @@
+package use_case;
+
+import java.util.ArrayList;
+
+public interface OrderDataAccess {
+
+        ArrayList<Order> read();
+
+        void update(ArrayList<Order> orders);
+}
